Add files via upload
This python script takes a list of gene symbols (formatted as a single column in a tab-delimited text file) and reports back the amount of citations this gene has in PubMed. Will overestimate citations for a gene when it is a common word, but it reports this with a warning. The script is run like this: python hype_my_science.py /path/to/genelist.txt